Move classes para os pacotes adequados
